Grand Finale! \n - new ckeditor.js module/class. \n - checks the body upon load for a ckeditor instance [data-ckeditor] \n - calls new CkeditorClass(item) to instantiate a new instance \n
This file is a module that is used to dynamically load the ckeditor.js
file when a new instance is detected, then proceeds to instantiate,
the new CKEDITOR instance. We do not include the ckeditor scripts via,
require or an import statement here because it will load the ckeditor
too soon before the required global variable is set. This method,
ensures we only load CKEDITOR when it is absolutely required, and keeps,
the bundling of the library to a minimum.